字段数组生成实体类的代码独立抽象出来
