+++ parser/src/main/kotlin/me/danwi/sqlex/parser/generate/GeneratedMethodFile.kt
@@ -144,7 +144,7 @@ class GeneratedMethodFile(
         } else {
             //如果是多列,则生成对应的实体类
             val resultClassName = method.returnType()?.text ?: "${methodName.pascalName}Result"
-            val resultClassSpec = generateResultClass(resultClassName, fields)
+            val resultClassSpec = fields.toEntityClass(resultClassName)
             //把实体类添加到内部类
             innerClasses.add(resultClassSpec)
             ClassName.get(packageName, className, resultClassName)
@@ -249,29 +249,6 @@ class GeneratedMethodFile(
         return methodSpec.build()
     }
 
-    private fun generateResultClass(resultClassName: String, fields: Array<Field>): TypeSpec {
-        val typeSpecBuilder = TypeSpec.classBuilder(resultClassName)
-            .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-        //判断列名是否重复
-        val duplicateFieldNames =
-            fields.groupBy(keySelector = { it.name.pascalName }, valueTransform = { it.name })
-                .filter { it.value.size > 1 }.values
-        if (duplicateFieldNames.isNotEmpty()) {
-            throw Exception("重复的列名 ${duplicateFieldNames.joinToString(", ")}")
-        }
-        //判断列名是否非法
-        val regex = ColumnNameRegex.ColumnNameRegex.toRegex()
-        val invalidFieldNames = fields.filter { !regex.matches(it.name.pascalName) }
-        if (invalidFieldNames.isNotEmpty()) {
-            throw Exception("非法的列名 ${invalidFieldNames.joinToString(", ") { it.name }}")
-        }
-        //给实体添加getter/setter
-        fields
-            .map { Pair(it.name, it.JavaType) }
-            .forEach { typeSpecBuilder.addColumnGetterAndSetter(it.first, it.second) }
-        return typeSpecBuilder.build()
-    }
-
     private fun generateAnnotation(
         paramList: SqlExMethodLanguageParser.ParamListContext?,
         namedParameterSQL: NamedParameterSQL,
@@ -371,27 +348,4 @@ class GeneratedMethodFile(
         //返回参数
         return parametersInMethod.map { ParameterSpec.builder(it.second, it.first).build() }
     }
-}
-
-//给实体类添加数据列getter/setter
-fun TypeSpec.Builder.addColumnGetterAndSetter(columnName: String, type: TypeName): TypeSpec.Builder {
-    this.addField(type, "_${columnName.pascalName}", Modifier.PRIVATE)
-    this.addMethod(
-        MethodSpec.methodBuilder("get${columnName.pascalName}")
-            .addModifiers(Modifier.PUBLIC)
-            .addStatement("return this._${columnName.pascalName}")
-            .returns(type)
-            .build()
-    )
-    this.addMethod(
-        MethodSpec.methodBuilder("set${columnName.pascalName}")
-            .addAnnotation(
-                AnnotationSpec.builder(SqlExColumnName::class.java).addMember("value", "\$S", columnName).build()
-            )
-            .addModifiers(Modifier.PUBLIC)
-            .addParameter(type, "value")
-            .addStatement("this._${columnName.pascalName} = value")
-            .build()
-    )
-    return this
 }
